match: /Od slot-only census closes 26 functions tree-wide
The brief was seven narrow accessors in playerData.h. Three are byte-proven:
CurrentTown (i8 spills a temp where retail has movsbl 0x45(%edx) with no slot),
Color (SetupHeroView 94.86 -> 95.30 in isolation, HERO's spill census 3/2 ->
2/2) and AvailableHeroId (CastleHandler 22.34 -> 25.53, SetupCastle
97.16 -> 98.21, Castle spill delta +1 -> 0).

The other four are NOT evidenced and are kept only for family coherence.
TownId lost its last call site to a permuted winner; HeroCount, TownCount and
HeroId have none at all. Flipping just those three produced byte-identical
objects tree-wide, so no measurement can decide them - a return-type edit to an
uncalled inline accessor does not reroll /Od parity on VC6. Revert them if the
header should carry only evidenced claims.

The premise also does not generalise: once an accessor returns i32 its /Ob1
expansion is byte-identical to a direct member read, so exactly one call site
wanted the accessor back. That site founds a rule - with one inlined call in a
binary operator, the call-bearing side is evaluated LAST - and all three
call-free spellings evaluate the other operand first.

The larger result came from noticing SetupCastle was instruction-identical to
retail with only its frame slots wrong. The MSVC 4.2 /Od name-hash model
predicts our VC6 layout exactly, so the lane scanned the whole tree for that
class: 32 functions byte-identical apart from slot names, 26 of which close
here. Among them PowEffect (4567 bytes), InitMap, REQUEST::Open,
swapManager::Open, DDCreateSurface, InitNetHost, InitNetGuest, SendRemoteData,
GetMap, UpdateArmies, WGUpdatePalette.

Two data-model corrections fell out of the frames. SAVED_MAP_NAME_CAPACITY is
16, not 44 - gMapName is char[16] and retail's savedName base address proves it
- with a declared-but-unreferenced 28-byte local above it that the 44 was
papering over. SendRemoteData's buffer is [REMOTE_MESSAGE_SIZE] plus a separate
unreferenced i32. Both slots are byte-proven; the inferred names say so.

EVENTS.cpp keeps master's version: this lane's base predated the EVENTS batch,
and its copy would have reverted PasswordEvent's CP1251 text to English and
renamed locals for functions master already has at 100%.

Tree 1388 -> 1414 exact, code bytes 53.39% -> 56.26%. No regressions.

+++ include/SOURCE/playerData.h
@@ -81,25 +81,25 @@ public:
     i32 CurrentHero(void) {
         return m_currentHero;
     }
-    i8 CurrentTown(void) {
+    i32 CurrentTown(void) {
         return m_currentTown;
     }
-    i8 Color(void) {
+    i32 Color(void) {
         return m_color;
     }
-    i8 HeroCount(void) {
+    i32 HeroCount(void) {
         return m_heroCount;
     }
-    i8 TownCount(void) {
+    i32 TownCount(void) {
         return m_townCount;
     }
-    i8 HeroId(i32 index) {
+    i32 HeroId(i32 index) {
         return m_heroIds[index];
     }
-    i8 TownId(i32 index) {
+    i32 TownId(i32 index) {
         return m_townIds[index];
     }
-    i8 AvailableHeroId(i32 index) {
+    i32 AvailableHeroId(i32 index) {
         return m_availableHeroIds[index];
     }
 };
